ai-agents/requirements.txt
litellm 
tavily-python
langgraph 
langchain_community 
langchain-google-genai 
langchain-groq
python-dotenv
colorama